add new file in spring boot practice

+++ Spring_MVC_blog/src/main/java/com/benson/springbootpractice/dao/CityDao.java
@@ -10,9 +10,9 @@ public interface CityDao {
     List<City> findAllCityFromDB();
     List<City> findAllFromDB();
 
-    City findCityByIdFromDB(Long id);
+    City findCityByIdFromDB(int id);
     void updateByIdFromDB(City city);
     void addCityFromDB(City city);
-    void deleteByIdFromDB(Long id);
+    void deleteByIdFromDB(int id);
 
 }
